+++ README.md
@@ -16,4 +16,4 @@ You are given movies loaded from the API and initial markup. Your task is to:
 - Implement a solution following the [React task guideline](https://github.com/mate-academy/react_task-guideline#react-tasks-guideline).
 - Use the [React TypeScript cheat sheet](https://mate-academy.github.io/fe-program/js/extra/react-typescript).
 - Open one more terminal and run tests with `npm test` to ensure your solution is correct.
-- Replace `<your_account>` with your Github username in the [DEMO LINK](https://<your_account>.github.io/react_movies-list/) and add it to the PR description.
+- Replace `<your_account>` with your Github username in the [DEMO LINK](https://OlyaNikora.github.io/react_movies-list/) and add it to the PR description.
