+++ src/components/types/Movie.ts
@@ -1,0 +1,7 @@
+export interface Movie {
+  title: string;
+  description: string;
+  imgUrl: string;
+  imdbUrl: string;
+  imdbId: string;
+}
