--- src/types/Movie.ts
@@ -1,5 +1,0 @@
-export interface Movie {
-  title: string;
-
-  // add all the other field
-}
